requirements.txt
python-dateutil>=2.8.0
requests>=2.8.0
# pandas>=1.1.0,<2.0.0
pandas>=1.1.0,<3.0.0
numpy>=1.19.0,<=1.26.4
scikit-learn>=1.3.0
pydantic>1.0.0,<3.0.0
fastparquet>=0.8.1
python-json-logger>=3.3.0
catboost>=1.0.3
# lightgbm>=3.3.2
pyjwt>=2.8.0
xhtml2pdf>=0.2.11,<0.3.0
python-bidi==0.4.2
ipywidgets>=8.1.0
# autofe
jarowinkler>=2.0.0
levenshtein>=0.25.1
psutil>=6.0.0
python-dotenv>=1.0.0

black
flake8
flake8-bugbear 
pytest
coverage
pytest-cov
pytest-datafiles
pytest-timeout
requests-mock
unittest-xml-reporting
pytest-parallel
py
build
twine
pytest-xdist